requirements.txt
Django~=3.2.13
django-bootstrap-v5
Pillow==9.1.1
face-recognition==1.3.0
gunicorn==20.1.0
cmake==3.22.4
psycopg2==2.9.3